pyproject.toml
[project]
name = "scrapper-function"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bs4>=0.0.2",
    "dotenv>=0.9.9",
    "fastapi>=0.116.1",
    "google-cloud-firestore>=2.21.0",
    "playwright>=1.55.0",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
]
